XP: Adjust formula calc to use floats

+++ src/game/Tools/Formulas.h
@@ -130,9 +130,9 @@ namespace MaNGOS
             return 17;
         }
 
-        inline uint32 BaseGain(uint32 unit_level, uint32 mob_level, ContentLevels content)
+        inline float BaseGain(uint32 unit_level, uint32 mob_level, ContentLevels content)
         {
-            uint32 nBaseExp = unit_level * 5;
+            float nBaseExp = unit_level * 5;
             switch (content)
             {
                 case CONTENT_1_60:  nBaseExp += 45;  break;
